refactor(sanity): add reusable perspective list hook
